requirements.txt
httpx
jmespath
pytesseract
Pillow